tests/wpt/web-platform-tests/app-history/app-history-entry/after-detach.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  let i_appHistory = i.contentWindow.appHistory;

  await i_appHistory.navigate("#1").finished;
  await i_appHistory.navigate("#2").finished;
  await i_appHistory.back().finished;

  assert_not_equals(i_appHistory, null);
  assert_not_equals(i_appHistory.current, null);
  assert_equals(i_appHistory.entries().length, 3);
  assert_true(i_appHistory.canGoBack, "canGoBack");
  assert_true(i_appHistory.canGoForward, "canGoForward");

  i.remove();

  assert_equals(i_appHistory.current, null);
  assert_equals(i_appHistory.entries().length, 0);
  assert_false(i_appHistory.canGoBack);
  assert_false(i_appHistory.canGoForward);
}, "appHistory.current/entries()/canGoBack/canGoForward after iframe removal");
</script>
